+++ CHANGELOG.md
@@ -11,6 +11,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Breaking change: Remove `FormRow` component and `k-FormRow` styles.
 - Breaking change: Remove `Search` component and `k-SearchInput` styles.
 - Breaking change: Convert `LinkList` component and delete `k-LinkList` styles.
+- Remove `k-Slider` and `k-SliderTooltip` styles.
+- Remove deprecated `Slider` components, in favor of `RangeSlider`:
+  - `Slider` ;
+  - `SliderBar` ;
+  - `SliderTooltip` ;
+  - `SliderWithTooltipAndPower`.
+- Remove `sliderKeyDownHandler`.
+- Remove `LoanSimulator` component and `k-LoanSimulator` styles.
 
 ## [3.0.0-beta.2] - 2021-02-15